i try to make it died with the write time

+++ philo/philo.h
@@ -6,7 +6,7 @@
 /*   By: acennadi <acennadi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/29 15:19:16 by acennadi          #+#    #+#             */
-/*   Updated: 2025/08/10 18:21:37 by acennadi         ###   ########.fr       */
+/*   Updated: 2025/08/10 20:01:32 by acennadi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -40,7 +40,7 @@ typedef struct s_philos
 	pthread_mutex_t			*left_fork;
 	pthread_mutex_t			*right_fork;
 	long long				last_meal;
-	int						eat_count;
+	long						eat_count;
 	struct s_configuration	*config;
 }							t_phios;
